Create abstract Argon2id interface
Don't handle Argon2id in LoginFacade.
